moved files from circuit to assertions folder
