+++ qiskit/assertions/assertmanager.py
