+++ qiskit/assertions/asserts.py
